+++ pkg/crypto/symmetric/symmetric.go
@@ -3,7 +3,6 @@ package symmetric
 import (
 	"crypto/aes"
 	"crypto/cipher"
-	"fmt"
 
 	"github.com/number571/go-peer/pkg/crypto/random"
 )
@@ -19,28 +18,27 @@ const (
 )
 
 type sAESCipher struct {
-	fKey []byte
+	fBlock cipher.Block
 }
 
 func NewAESCipher(pKey []byte) ICipher {
 	if len(pKey) != CAESKeySize {
 		panic("len(pKey) != CAESKeySize")
 	}
+	block, err := aes.NewCipher(pKey)
+	if err != nil {
+		return nil
+	}
 	return &sAESCipher{
-		fKey: pKey,
+		fBlock: block,
 	}
 }
 
 func (p *sAESCipher) EncryptBytes(pMsg []byte) []byte {
-	block, err := aes.NewCipher(p.fKey)
-	if err != nil {
-		return nil
-	}
-
-	blockSize := block.BlockSize()
+	blockSize := p.fBlock.BlockSize()
 	iv := random.NewStdPRNG().GetBytes(uint64(blockSize))
 
-	stream := cipher.NewCFBEncrypter(block, iv)
+	stream := cipher.NewCFBEncrypter(p.fBlock, iv)
 	result := make([]byte, len(pMsg)+len(iv))
 	copy(result[:blockSize], iv)
 
@@ -49,29 +47,16 @@ func (p *sAESCipher) EncryptBytes(pMsg []byte) []byte {
 }
 
 func (p *sAESCipher) DecryptBytes(pMsg []byte) []byte {
-	block, err := aes.NewCipher(p.fKey)
-	if err != nil {
-		return nil
-	}
-
-	blockSize := block.BlockSize()
+	blockSize := p.fBlock.BlockSize()
 	if len(pMsg) < blockSize {
 		return nil
 	}
 
-	stream := cipher.NewCFBDecrypter(block, pMsg[:blockSize])
+	stream := cipher.NewCFBDecrypter(p.fBlock, pMsg[:blockSize])
 	result := make([]byte, len(pMsg)-blockSize)
 
 	stream.XORKeyStream(result, pMsg[blockSize:])
 	return result
-}
-
-func (p *sAESCipher) ToString() string {
-	return fmt.Sprintf("Key(%s){%X}", p.GetType(), p.ToBytes())
-}
-
-func (p *sAESCipher) ToBytes() []byte {
-	return p.fKey
 }
 
 func (p *sAESCipher) GetType() string {