+++ pkg/storage/database/encrypt.go
@@ -8,37 +8,40 @@ import (
 	"github.com/number571/go-peer/pkg/errors"
 )
 
-func doEncrypt(pCipher symmetric.ICipher, pDataBytes []byte) []byte {
+func doEncrypt(pCipher symmetric.ICipher, pAuthKey []byte, pDataBytes []byte) []byte {
+	encDataBytes := pCipher.EncryptBytes(pDataBytes)
 	return bytes.Join(
 		[][]byte{
 			hashing.NewHMACSHA256Hasher(
-				pCipher.ToBytes(),
-				pDataBytes,
+				pAuthKey,
+				encDataBytes,
 			).ToBytes(),
-			pCipher.EncryptBytes(pDataBytes),
+			encDataBytes,
 		},
 		[]byte{},
 	)
 }
 
-func tryDecrypt(pCipher symmetric.ICipher, pEncBytes []byte) ([]byte, error) {
+func tryDecrypt(pCipher symmetric.ICipher, pAuthKey []byte, pEncBytes []byte) ([]byte, error) {
 	if len(pEncBytes) < hashing.CSHA256Size+symmetric.CAESBlockSize {
 		return nil, errors.NewError("incorrect size of encrypted data")
 	}
 
-	decBytes := pCipher.DecryptBytes(pEncBytes[hashing.CSHA256Size:])
-	if decBytes == nil {
-		return nil, errors.NewError("failed decrypt message")
-	}
+	encDataBytes := pEncBytes[hashing.CSHA256Size:]
 
 	gotHashed := pEncBytes[:hashing.CSHA256Size]
 	newHashed := hashing.NewHMACSHA256Hasher(
-		pCipher.ToBytes(),
-		decBytes,
+		pAuthKey,
+		encDataBytes,
 	).ToBytes()
 
 	if !bytes.Equal(gotHashed, newHashed) {
 		return nil, errors.NewError("incorrect hash of decrypted data")
+	}
+
+	decBytes := pCipher.DecryptBytes(encDataBytes)
+	if decBytes == nil {
+		return nil, errors.NewError("failed decrypt message")
 	}
 
 	return decBytes, nil